pyproject.toml
[tool.poetry]
name = "marketing"
version = "0.1.0"
description = ""
authors = ["kuribayashi <tktkchestnut@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
bs4 = "^0.0.1"
python-pptx = "^0.6.19"
selenium = "^3.141.0"
chromedriver-binary = "^92.0.4515.107"
pandas = "^1.3.1"
pysen = "^0.9.1"
isort = "^5.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.9"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
line_length = 130
py_version = "py37"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]